🚚 Move party size specific components to own folder
