+++ src/routes/runs/dfa/[party=partySize]/DfaPartyModalRunFilters.svelte
@@ -2,7 +2,7 @@
 	import Divider from '$lib/Components/Divider.svelte';
 	import ServerFilter from '$lib/Components/Filters/ServerFilter.svelte';
 	import Modal from '$lib/Components/Modal.svelte';
-	import DfaSupportFilter from './DfaSupportFilter.svelte';
+	import DfaSupportFilter from '../DfaSupportFilter.svelte';
 	import { createEventDispatcher } from 'svelte';
 	import { t } from 'svelte-i18n';
 	const dispatch = createEventDispatcher();
