+++ src/routes/runs/dfa/solo/DfaSoloRunsTable.svelte
