+++ src/routes/runs/purple/solo/PurpleSoloRunFilters.svelte
@@ -1,22 +1,23 @@
 <script lang="ts">
 	import Divider from '$lib/Components/Divider.svelte';
-	import PurpleRules from './PurpleRules.svelte';
-	import PurplePartyModalRunFilters from './PurplePartyModalRunFilters.svelte';
-	import PurpleCategorySelector from './PurpleCategorySelector.svelte';
-	import { partyRunFilters, type PurplePartySearchFilters } from './purpleRunFilterStore';
+	import PurpleRules from '../PurpleRules.svelte';
+	import PurpleSoloModalRunFilters from './PurpleSoloModalRunFilters.svelte';
+	import PurpleCategorySelector from '../PurpleCategorySelector.svelte';
+	import { soloRunFilters, type PurpleSoloSearchFilters } from '../purpleRunFilterStore';
 	import { t } from 'svelte-i18n';
 
-	let filters: PurplePartySearchFilters = {
+	let filters: PurpleSoloSearchFilters = {
+		server: 'no_filter',
 		region: 'stia',
 		rank: '1',
-		server: 'no_filter'
+		class: 'no_filter'
 	};
 
 	const applyFilters = () => {
-		partyRunFilters.set({ ...filters });
+		soloRunFilters.set({ ...filters });
 	};
 
-	partyRunFilters.subscribe((f) => {
+	soloRunFilters.subscribe((f) => {
 		filters = { ...f };
 	});
 </script>
@@ -33,7 +34,11 @@
 	<Divider class="-mx-1 my-0" />
 	<div class="flex flex-row flex-wrap place-content-center items-stretch">
 		<div class="m-1 md:flex-1">
-			<PurplePartyModalRunFilters bind:server={filters.server} on:applyFilters={applyFilters} />
+			<PurpleSoloModalRunFilters
+				bind:server={filters.server}
+				bind:mainClass={filters.class}
+				on:applyFilters={applyFilters}
+			/>
 		</div>
 		<div class="m-1 md:flex-initial">
 			<PurpleRules />
